# swagger_schema.json
{"openapi":"3.0.0","info":{"version":"1.0.0","title":"Hash linked log","description":"Hash linked log"},"servers":[{"url":"/","description":"Local Dev"}],"paths":{"/logs/entry":{"post":{"tags":["Hash linked log"],"description":"Attempts to write an entry to the end of the hash linked log","operationId":"writeLogEntry","parameters":[{"in":"body","name":"message","schema":{"type":"string","minLength":1,"maxLength":200},"required":true}],"responses":{"204":{"description":"The entry was written ok"},"400":{"content":{"application/json":{"schema":{"type":"object","required":["internal_code"],"properties":{"errors":{"type":"array","items":{"type":"object","required":["message"],"properties":{"message":{"type":"string","example":"Some descriptive error message"},"code":{"type":"string","example":"A22"}}},"example":[]},"internal_code":{"type":"string","enum":["invalid_log_message","entry_was_out_paced_by_another_one"]}}}}}},"500":{"content":{"application/json":{"schema":{"type":"object","required":["internal_code"],"properties":{"errors":{"type":"array","items":{"type":"object","required":["message"],"properties":{"message":{"type":"string","example":"Some descriptive error message"},"code":{"type":"string","example":"A22"}}},"example":[]},"internal_code":{"type":"string","enum":["default_error","internal_error","corrupted_log_file"]}}}}}},"503":{"content":{"application/json":{"schema":{"type":"object","required":["internal_code"],"properties":{"errors":{"type":"array","items":{"type":"object","required":["message"],"properties":{"message":{"type":"string","example":"Some descriptive error message"},"code":{"type":"string","example":"A22"}}},"example":[]},"internal_code":{"type":"string","enum":["service_unavailable"]}}}}}}}}},"/logs":{"get":{"tags":["Hash linked log"],"description":"Get the logs","operationId":"getLogs","responses":{"200":{"description":"The logs","content":{"application/json":{"schema":{"type":"object","properties":{"logs":{"type":"array","example":[["2020-05-31T12:19:34.231Z","Some message"],["2020-06-31T03:11:01.643Z","Some message"],["2020-02-31T02:02:12.8122Z","Some message"]]}}}}}},"500":{"content":{"application/json":{"schema":{"type":"object","required":["internal_code"],"properties":{"errors":{"type":"array","items":{"type":"object","required":["message"],"properties":{"message":{"type":"string","example":"Some descriptive error message"},"code":{"type":"string","example":"A22"}}},"example":[]},"internal_code":{"type":"string","enum":["default_error","internal_error","corrupted_log_file"]}}}}}},"503":{"content":{"application/json":{"schema":{"type":"object","required":["internal_code"],"properties":{"errors":{"type":"array","items":{"type":"object","required":["message"],"properties":{"message":{"type":"string","example":"Some descriptive error message"},"code":{"type":"string","example":"A22"}}},"example":[]},"internal_code":{"type":"string","enum":["service_unavailable"]}}}}}}}}}},"components":{"schemas":{},"securitySchemes":{}}}